Remove legacy IE8-9 tags
This might be a disputed commit so here's my reasoning:
- Not only has IE9 been EoL for 4 years, the entire Internet Explorer
brand has been dropped by Microsoft as of 2020.
- <IE9 accounts for less than 0.2% of global traffic

Some might say "why remove it, it doesn't do any harm",
which I disagree with. This template is included in any project
we build so it should be as clean as possible.