Move the toasts wrapper div inside the notification template
I noticed we only wrapped the notifications include with the new Vue toast div in base.html.twig,
but this include is also used on user.html.twig and empty.html.twig. So the new toasts didn't work
on any template extending those, for example login pages, or register pages. I could've added
the div with the correct ID in those templates as well, but instead I moved the wrapper div
into the include, so all templates that use it will have the correct markup.