Improve the base DTO class
This iteration will handle recursion, so nested DTOs will
not be a problem anymore.

It also works with the constructor instead of the static
 `from` method call. The only downside to this approach is
that the source objetc that is passed to the DTO isn't
properly type-hinted when you fetch it using the getSource
method. You need to provide an inline @var docblock before
your IDE will autocomplete it and PHPStan will accept it.