Append our default classes on menu items instead of just setting them
Before, the necessary classes were simply set on our menu items. This meant that even
if you set a class on a specific item in your PHP code: $item->setAttribute('class', 'foo')'
it would not show up in he template, because it would be overwritten.

We should append the required classes on top of the ones we set in our MenuListener.