Fix nested form collection types
For nested form collection types to work, we need to set different prototype name. In this example:  field___name___sub___name__, both collection type have the same prototype name and will both be replaced when adding an item in the first collection type.

To make field___name___sub___name2__ work, we need to read the prototype name from the dataset when replacing it in JavaScript.